requirements.txt
langchain
openai
tiktoken
flask
python-dotenv